Support year in netcoupe crawler
Netcoupe partitions flights into different websites (different
endpoints) based on the year, with the prefix 'archive<year>', and 'www'
being the current year.

Take this logic into account when instantiating a new Netcoupe crawler
by taking an year param on the constructor, and using it to calculate
the endpoint to be contacted.

As it is today, netcoupe only keeps the 3 last years of archive
available.